v0.1/requirements.txt
pandas
yfinance
lxml
beautifulsoup4
google-cloud-bigquery
flask
gunicorn